+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+# 0.23.1
+## No Breaking Changes
+
+## Features Added
+  - Extended `RobotoPrincipal` to include devices + invocations, and added methods to convert to and from a canonical string format.
+  - Add `RobotoSearch::for_roboto_client` and `RobotoClient.for_profile` to simplify code snippets for users with multiple profiles.
+
+## Bugs Fixed
+  - 
+
 # 0.23.0
 ## No Breaking Changes